+++ dev-resources/org/example/db.clj
@@ -8,43 +8,43 @@
        [{:id "1000"
          :name "Luke Skywalker"
          :friends ["1002", "1003", "2000", "2001"]
-         :appears-in ["NEWHOPE" "EMPIRE" "JEDI"]
-         :home-planet "Tatooine"
-         :force-side "3001"}
+         :appears_in ["NEWHOPE" "EMPIRE" "JEDI"]
+         :home_planet "Tatooine"
+         :force_side "3001"}
         {:id "1001"
          :name "Darth Vader"
          :friends ["1004"]
-         :appears-in ["NEWHOPE" "EMPIRE" "JEDI"]
-         :home-planet "Tatooine"
-         :force-side "3000"}
+         :appears_in ["NEWHOPE" "EMPIRE" "JEDI"]
+         :home_planet "Tatooine"
+         :force_side "3000"}
         {:id "1003"
          :name "Leia Organa"
          :friends ["1000", "1002", "2000", "2001"]
-         :appears-in ["NEWHOPE" "EMPIRE" "JEDI"]
-         :home-planet "Alderaan"
-         :force-side "3001"}
+         :appears_in ["NEWHOPE" "EMPIRE" "JEDI"]
+         :home_planet "Alderaan"
+         :force_side "3001"}
         {:id "1002"
          :name "Han Solo"
          :friends ["1000", "1003", "2001"]
-         :appears-in ["NEWHOPE" "EMPIRE" "JEDI"]
-         :force-side "3001"}
+         :appears_in ["NEWHOPE" "EMPIRE" "JEDI"]
+         :force_side "3001"}
         {:id "1004"
          :name "Wilhuff Tarkin"
          :friends ["1001"]
-         :appears-in ["NEWHOPE"]
-         :force-side "3000"}]))
+         :appears_in ["NEWHOPE"]
+         :force_side "3000"}]))
 
 (def ^:private droids-data
   (map #(schema/tag-with-type % :droid)
        [{:id "2001"
          :name "R2-D2"
          :friends ["1000", "1002", "1003"]
-         :appears-in ["NEWHOPE" "EMPIRE" "JEDI"]
+         :appears_in ["NEWHOPE" "EMPIRE" "JEDI"]
          :primary-function "ASTROMECH"}
         {:id "2000"
          :name "C-3PO"
          :friends ["1000", "1002", "1003", "2001"]
-         :appears-in ["NEWHOPE" "EMPIRE" "JEDI"]
+         :appears_in ["NEWHOPE" "EMPIRE" "JEDI"]
          :primary-function "PROTOCOL"}]))
 
 (def ^:private character-data (concat humans-data droids-data))